make updates to make sure the method has the right parameter type as well as call the method for the PlayerEntry property in the WeeklyLeaderboardEntriesResponse() class

+++ Object/LeaderboardEntry.cs
@@ -32,7 +32,7 @@ namespace spikewall.Object
         public ulong? league { get; set; }
         public ulong? maxScore { get; set; }
 
-        public static LeaderboardEntry PlayerToLeaderboardEntry(Player player, int mode)
+        public static LeaderboardEntry PlayerToLeaderboardEntry(Player player, long mode)
         {
             var friendID = player.ID;
             var name = player.Username;
